Add MarkdownView component
* Rename MarkdownView to RemoteMarkdown
* Create MarkdownView that makes use of RemoteMarkdown
* Implement about view
* Make use of MarkdownView in rules view
