+++ src/client/views/about/view.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import Helmet from 'react-helmet';
 
+import MarkdownView from '../../components/markdown-view';
+
+const ABOUT_URL = 'https://raw.githubusercontent.com/TF2PickupNET/Info/master/ABOUT.md';
+
 /**
  * The view for the about page.
  *
@@ -12,6 +16,8 @@ export default function View() {
       <Helmet>
         <title>About</title>
       </Helmet>
+
+      <MarkdownView url={ABOUT_URL} />
     </div>
   );
 }
