+++ src/client/views/rules/view.jsx
@@ -1,9 +1,5 @@
-import React, { PureComponent } from 'react';
-import PropTypes from 'prop-types';
+import React from 'react';
 import Helmet from 'react-helmet';
-import injectSheet from 'react-jss';
-import Aux from 'react-aux';
-import { Card } from 'materialize-react';
 
 import MarkdownView from '../../components/markdown-view';
 
@@ -12,43 +8,16 @@ const RULES_URL = 'https://raw.githubusercontent.com/TF2PickupNET/Info/master/RU
 /**
  * The view for the rules page.
  *
- * @class
+ * @returns {JSX} - Returns the view.
  */
-class View extends PureComponent {
-  static propTypes = {
-    classes: PropTypes.shape({
-      container: PropTypes.string.isRequired,
-      rules: PropTypes.string.isRequired,
-    }).isRequired,
-  };
+export default function View() {
+  return (
+    <div>
+      <Helmet>
+        <title>Rules</title>
+      </Helmet>
 
-  static styles = {
-    container: {
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'center',
-    },
-
-    rules: {
-      width: '100%',
-      maxWidth: 1000,
-    },
-  }
-
-  render() {
-    return (
-      <Aux>
-        <Helmet>
-          <title>Rules</title>
-        </Helmet>
-        <div className={this.props.classes.container}>
-          <Card className={this.props.classes.rules}>
-            <MarkdownView url={RULES_URL} />
-          </Card>
-        </div>
-      </Aux>
-    );
-  }
+      <MarkdownView url={RULES_URL} />
+    </div>
+  );
 }
-
-export default injectSheet(View.styles)(View);
